requirements.txt
boto3
llama-index
llama-index-llms-bedrock
llama-index-embeddings-bedrock
llama-index-vector-stores-tidbvector
llama-index-readers-web
streamlit
python-dotenv